# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "ChangeEmailRequest": {
        "additionalProperties": false,
        "properties": {
          "current_password": {
            "title": "Current Password",
            "type": "string"
          },
          "new_email": {
            "format": "email",
            "title": "New Email",
            "type": "string"
          }
        },
        "required": [
          "new_email",
          "current_password"
        ],
        "title": "ChangeEmailRequest",
        "type": "object"
      },
      "ChangePasswordRequest": {
        "additionalProperties": false,
        "properties": {
          "current_password": {
            "title": "Current Password",
            "type": "string"
          },
          "new_password": {
            "maxLength": 72,
            "minLength": 8,
            "title": "New Password",
            "type": "string"
          }
        },
        "required": [
          "current_password",
          "new_password"
        ],
        "title": "ChangePasswordRequest",
        "type": "object"
      },
...
      "ConfirmEmailChangeRequest": {
        "additionalProperties": false,
        "properties": {
          "token": {
            "maxLength": 512,
            "minLength": 1,
            "title": "Token",
            "type": "string"
          }
        },
        "required": [
          "token"
        ],
        "title": "ConfirmEmailChangeRequest",
        "type": "object"
      },
...
      "DeleteAccountRequest": {
        "additionalProperties": false,
        "properties": {
          "confirm": {
            "default": false,
            "title": "Confirm",
            "type": "boolean"
          },
          "password": {
            "title": "Password",
            "type": "string"
          }
        },
        "required": [
          "password"
        ],
        "title": "DeleteAccountRequest",
        "type": "object"
      },
...
      "EditorPreferences": {
        "additionalProperties": false,
        "description": "Per-user editor settings; server clamps font size and enums the rest.",
        "properties": {
          "font_size": {
            "default": 14,
            "title": "Font Size",
            "type": "integer"
          },
          "keymap": {
            "default": "default",
            "enum": [
              "default",
              "vim",
              "emacs"
            ],
            "title": "Keymap",
            "type": "string"
          },
          "theme": {
            "default": "system",
            "enum": [
              "light",
              "dark",
              "system"
            ],
            "title": "Theme",
            "type": "string"
          }
        },
        "title": "EditorPreferences",
        "type": "object"
      },
      "ErrorBody": {
        "description": "The ``error`` object carried by every error response.",
        "properties": {
          "details": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "description": "Optional structured detail (e.g. per-field validation errors).",
            "title": "Details"
          },
          "message": {
            "description": "Human-readable summary of the error.",
            "title": "Message",
            "type": "string"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "description": "Correlation id of the request that produced the error.",
            "title": "Request Id"
          },
          "type": {
            "description": "Machine-readable, snake_case error code.",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "message"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
      "ErrorEnvelope": {
        "description": "Top-level envelope wrapping an :class:`ErrorBody`.",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MessageResponse": {
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          }
        },
        "required": [
          "detail"
        ],
        "title": "MessageResponse",
        "type": "object"
      },
...
      "UpdateProfileRequest": {
        "additionalProperties": false,
        "description": "Patch the profile; both fields optional (at least one should be sent).",
        "properties": {
          "avatar_url": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar Url"
          },
          "display_name": {
            "anyOf": [
              {
                "maxLength": 100,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          }
        },
        "title": "UpdateProfileRequest",
        "type": "object"
      },
...
      "UserMe": {
        "description": "The signed-in user's own view, including profile + preferences (spec 59).\n\nNOTE (spec 68 #22): ``GET /api/v1/users/me`` returns this model \u2014 a strict\nsuperset of the spec-08 ``UserPublic`` (adds ``avatar_url``,\n``editor_preferences``, ``pending_email``). This is intentional over-delivery\nintroduced by spec 59; the extra fields never break a ``UserPublic`` consumer.",
        "properties": {
          "avatar_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar Url"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "editor_preferences": {
            "$ref": "#/components/schemas/EditorPreferences",
            "default": {
              "font_size": 14,
              "keymap": "default",
              "theme": "system"
            }
          },
          "email": {
            "title": "Email",
            "type": "string"
          },
          "email_confirmed": {
            "title": "Email Confirmed",
            "type": "boolean"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "is_admin": {
            "title": "Is Admin",
            "type": "boolean"
          },
          "pending_email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pending Email"
          }
        },
        "required": [
          "id",
          "email",
          "display_name",
          "is_admin",
          "email_confirmed",
          "created_at"
        ],
        "title": "UserMe",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Inkstave backend API.",
    "title": "Inkstave API",
    "version": "0.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/users/confirm-email-change": {
      "post": {
        "operationId": "confirm_email_change_api_v1_users_confirm_email_change_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmEmailChangeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserMe"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Confirm an email change via its token (unauthenticated)",
        "tags": [
          "users"
        ]
      }
    },
    "/api/v1/users/me": {
      "delete": {
        "operationId": "delete_me_api_v1_users_me_delete",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteAccountRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete the account (hard delete; cascades owned projects)",
        "tags": [
          "users"
        ]
      },
      "get": {
        "operationId": "read_me_api_v1_users_me_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserMe"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "The authenticated user",
        "tags": [
          "users"
        ]
      },
      "patch": {
        "operationId": "update_me_api_v1_users_me_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfileRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserMe"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update profile",
        "tags": [
          "users"
        ]
      }
    },
    "/api/v1/users/me/change-email": {
      "post": {
        "operationId": "change_email_api_v1_users_me_change_email_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeEmailRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Start an email change (confirmation sent to the new address)",
        "tags": [
          "users"
        ]
      }
    },
    "/api/v1/users/me/change-password": {
      "post": {
        "operationId": "change_password_api_v1_users_me_change_password_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Change password (signs out other sessions)",
        "tags": [
          "users"
        ]
      }
    },
    "/api/v1/users/me/editor-preferences": {
      "put": {
        "operationId": "put_editor_preferences_api_v1_users_me_editor_preferences_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EditorPreferences"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EditorPreferences"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Set editor preferences",
        "tags": [
          "users"
        ]
      }
    },
...
  }
}